Moved benchmark client and server into cmd folder
