+++ benchmark/Makefile
@@ -5,9 +5,9 @@ GORUMS := benchmark_gorums.pb.go
 
 PROTOC_FLAGS := -I=../.:.
 
-.PHONY: all benchmain server
+.PHONY: all
 
-all: $(PROTO_GO) $(GRPC) $(GORUMS) benchmain server
+all: $(PROTO_GO) $(GRPC) $(GORUMS)
 
 $(PROTO_GO): %.pb.go : %.proto
 	@protoc --go_out=paths=source_relative:. $(PROTOC_FLAGS) $^
@@ -17,9 +17,3 @@ $(GRPC): %_grpc.pb.go : %.proto
 
 $(GORUMS): %_gorums.pb.go : %.proto
 	@protoc --gorums_out=paths=source_relative,trace=true:. $(PROTOC_FLAGS) $^
-
-benchmain:
-	@go build -o benchmain/benchmain ./benchmain
-
-server:
-	@go build -o server/server ./server
